No more clipping and forcing into walls
Tried to add accelerating platforms, but it actually made
things worse in terms of bouncing. I did get rid of reference velocites
based on moving platforms since friction handles that now. And I stopped
clipping by making the body box a little smaller and stopped forcing into
walls by keeping a wall count and not continuing motion when you are in
contact with a wall.